+++ package/R/powertbf01.R
@@ -20,7 +20,9 @@
 #'     \code{"two.sample"} (default), \code{"one.sample"}, or \code{"paired"}
 #' @param alternative Direction of the test. Can be either \code{"two.sided"}
 #'     (default), \code{"less"} , or \code{"greater"}. The latter two truncate
-#'     the analysis prior to negative and positive effects, respectively
+#'     the analysis prior to negative and positive effects, respectively. If set
+#'     to \code{"less"} or \code{"greater"}, the power is only computed based on
+#'     data with effect estimates in the direction of the alternative
 #' @param dpm Mean of the normal design prior assigned to the standardized mean
 #'     difference. Defaults to the analysis prior location
 #' @param dpsd Standard deviation of the normal design prior assigned to the