+++ package/R/ptbf01.R
@@ -2,7 +2,7 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
                     pscale = 1/sqrt(2), pdf = 1, dpm = plocation, dpsd = pscale,
                     type = c("two.sample", "one.sample", "paired"),
                     alternative = c("two.sided", "less", "greater"),
-                    lower.tail = TRUE, ...) {
+                    lower.tail = TRUE, drange = "adaptive", ...) {
     ## input checks
     stopifnot(
         length(k) == 1,
@@ -49,7 +49,11 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
 
         length(lower.tail) == 1,
         is.logical(lower.tail),
-        !is.na(lower.tail)
+        !is.na(lower.tail),
+
+        (is.numeric(drange) && length(drange) == 2 && all(is.finite(drange)) &&
+         drange[2] > drange[1]) || (is.character(drange) && length(drange) == 1 &&
+                                    !is.na(drange) && drange == "adaptive")
     )
     type <- match.arg(type)
     alternative <- match.arg(alternative)
@@ -68,25 +72,22 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
         neff <- n1
     }
 
-
-    ## HACK this is quite a hack, improve! perhaps implement an argument whether
-    ## the search range should be determined adpatively or not
-
-    ## determine effect
-    ## estimate region where BF < k for specified sample size
+    ## determine effect estimate region where BF < k for specified sample size
     se <- 1/sqrt(neff) # standard error of SMD assuming variance is known
     estsd <- sqrt(se^2 + dpsd^2) # standard deviation of SMD under design prior
     rootFun <- function(est) {
         tbf01(t = (est - null)/se, n1 = n1, n2 = n2, plocation = plocation,
               pscale = pscale, pdf = pdf, type = type,
-              alternative = alternative) - k
+              alternative = alternative, log = TRUE) - log(k)
     }
 
     if (alternative == "two.sided") {
         if (k > 1) {
             ## check whether BF > k > 1 is achievable for given sample size
             ## maximum BF is obtained when est = null
-            if (rootFun(null) < 0) {
+            maxBF <- exp(rootFun(null) + log(k))
+            if (is.nan(maxBF)) return(NaN)
+            if (maxBF < k) {
                 if (lower.tail == FALSE) {
                     return(0)
                 } else {
@@ -96,37 +97,58 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
         }
 
         ## guess search range based on search range from z-test BF
-        suppressWarnings({
-            X <- (log(1 + neff*pscale^2) + (null - plocation)^2/pscale^2 - log(k^2))*
-                (1 + 1/neff/pscale^2)/neff
-            if (is.nan(X)) X <- 0.1
-            M <- (-null - (null - plocation)/neff/pscale^2)
-            lowerz <- -sqrt(X) - M # lower limit z-test BF
-            upperz <- sqrt(X) - M # upper limit z-test BF
-        })
+        ## TODO improve robustness of adaptive strategy
+        if (!is.numeric(drange) && drange == "adaptive") {
+            suppressWarnings({
+                X <- (log(1 + neff*pscale^2) + (null - plocation)^2/pscale^2 - log(k^2))*
+                    (1 + 1/neff/pscale^2)/neff
+                if (is.nan(X)) X <- 0.1
+                M <- (-null - (null - plocation)/neff/pscale^2)
+                critz1 <- -sqrt(X) - M # limit 1 z-test BF
+                critz2 <- sqrt(X) - M # limit 2 z-test BF
+            })
 
-        ## search for critical values
-        lowert <- (lowerz - 0.1)*6 # add a bit and blow up by factor 6
-        uppert <- (upperz + 0.1)*6 # add a bit and blow up by factor 6
-        upper <- try(stats::uniroot(f = rootFun, interval = c(0, abs(uppert)),
-                                    ... = ...)$root,
+            ## search for critical values
+            meancritz <- (critz1 + critz2)/2
+            if (critz1 < critz2) {
+                ## searchIntLow <- c(5*(critz1 - 0.01), meancritz)
+                ## searchIntUp <- c(meancritz, 5*(critz2 + 0.01))
+                searchIntLow <- c(critz1 - 0.01, meancritz)
+                searchIntUp <- c(meancritz, critz2 + 0.01)
+            } else {
+                ## searchIntLow <- c(5*(critz2 - 0.01), meancritz)
+                ## searchIntUp <- c(meancritz, 5*(critz1 + 0.01))
+                searchIntLow <- c(critz2 - 0.01, meancritz)
+                searchIntUp <- c(meancritz, critz1 + 0.01)
+            }
+        } else {
+            meant <- mean(drange)
+            searchIntLow <- c(drange[1], meant)
+            searchIntUp <- c(meant, drange[2])
+        }
+        upper <- try(stats::uniroot(f = rootFun, interval = searchIntUp,
+                                    extendInt = "yes", ... = ...)$root,
                      silent = TRUE)
-        lower <- try(stats::uniroot(f = rootFun, interval = c(-abs(lowert), 0),
-                                    ... = ...)$root,
+        lower <- try(stats::uniroot(f = rootFun, interval = searchIntLow,
+                                    extendInt = "yes", ... = ...)$root,
                      silent = TRUE)
 
         ## compute power
         if (inherits(upper, "try-error")) {
             powup <- 0
+            uperr <- TRUE
         } else {
             powup <- stats::pnorm(q = upper, mean = dpm, sd = estsd, lower.tail = FALSE)
+            uperr <- FALSE
         }
         if (inherits(lower, "try-error")) {
             powlow <- 0
+            lowerr <- TRUE
         } else {
             powlow <- stats::pnorm(q = lower, mean = dpm, sd = estsd, lower.tail = TRUE)
+            lowerr <- FALSE
         }
-        if (powup == 0 && powlow == 0) {
+        if ((uperr == TRUE) && (lowerr == TRUE)) {
             warning("Numerical problems finding critical value")
             pow <- NaN
         } else {
@@ -134,27 +156,39 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
         }
     } else {
         ## one-sided alternatives
-        ## search for critical value
-        searchRange <- dpm + c(-1, 1)*5*estsd # search around design prior
-        ## searchRange[1] <- pmin(searchRange[1], null)
-        ## searchRange[2] <- pmax(searchRange[2], null)
-        ## lowert <- pmin(lowerz, null) - 0.01
-        ## uppert <- pmax(upperz, null) + 0.01
-        ## searchRange <- c(lowert, uppert)
-        crit <- try(stats::uniroot(f = rootFun, interval = searchRange,
-                                   extendInt = "no", ... = ...)$root,
-                    silent = TRUE)
-        if (inherits(crit, "try-error")) {
-            ## power at the design prior mean
-            dpmpow <- rootFun(dpm)
-            if (is.nan(dpmpow)) {
-                warning("Numerical problems finding critical value")
-                pow <- NaN
-            } else if (dpmpow < 0) {
-                pow <- 1
+        if (k > 1) {
+            ## find maximum BF to see whether BF = k is possible
+            opt <- stats::optim(par = 0, fn = rootFun, control = list(fnscale = -1),
+                                method = "BFGS")
+            if (opt$convergence != 0) {
+                warning("numerical problems finding maximum BF")
+                return(NaN)
             } else {
-                pow <- 0
+                if (opt$value < 0) {
+                    ## maximum BF is smaller than k
+                    if (lower.tail == TRUE) {
+                        return(1)
+                    } else {
+                        return(0)
+                    }
+                }
             }
+        }
+
+        if (!is.numeric(drange) && drange == "adaptive") {
+            ## extend the search range if critical value not contained
+            searchRange <- c(-0.1, 0.1)
+            crit <- try(stats::uniroot(f = rootFun, interval = searchRange,
+                                       extendInt = "yes", ... = ...)$root,
+                        silent = TRUE)
+        } else {
+            crit <- try(stats::uniroot(f = rootFun, interval = drange,
+                                       extendInt = "no", ... = ...)$root,
+                        silent = TRUE)
+        }
+        if (inherits(crit, "try-error")) {
+            warning("Numerical problems finding critical value")
+            pow <- NaN
         } else {
             if (alternative == "greater") {
                 pow <- stats::pnorm(q = crit, mean = dpm, sd = estsd,
@@ -193,7 +227,10 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
 #' @param type Type of \eqn{t}-test associated with \eqn{t}-statistic. Can be
 #'     \code{"two.sample"} (default), \code{"one.sample"}, or \code{"paired"}
 #' @param alternative Direction of the test. Can be either \code{"two.sided"}
-#'     (default), \code{"less"}, or \code{"greater"}
+#'     (default), \code{"less"}, or \code{"greater"}. The latter two truncate
+#'     the analysis prior to negative and positive effects, respectively. If set
+#'     to \code{"less"} or \code{"greater"}, the power is only computed based on
+#'     data with effect estimates in the direction of the alternative
 #' @param dpm Mean of the normal design prior assigned to the standardized mean
 #'     difference. Defaults to the analysis prior location
 #' @param dpsd Standard deviation of the normal design prior assigned to the
@@ -204,6 +241,10 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
 #' @param lower.tail Logical indicating whether Pr(BF \eqn{\leq} \code{k})
 #'     (\code{TRUE}) or Pr(BF \eqn{>} \code{k}) (\code{FALSE}) should be
 #'     computed. Defaults to \code{TRUE}
+#' @param drange Standardized mean difference search range over which the
+#'     critical values are searched for. Can be either set to a numerical range
+#'     or to \code{"adaptive"} (default) which determines the range in an
+#'     adaptive way from the other input parameters
 #' @param ... Other arguments passed to \code{stats::uniroot}
 #'
 #' @return The probability that the Bayes factor is less or greater (depending
@@ -215,16 +256,16 @@ ptbf01. <- function(k = 1/10, n, n1 = n, n2 = n, null = 0, plocation = 0,
 #'
 #' @examples
 #' ## example from Schönbrodt and Wagenmakers (2018, p. 135)
-#' ptbf01(k = 1/6, n = 146, dpm = 0.5, dps = 0, alternative = "greater")
+#' ptbf01(k = 1/6, n = 146, dpm = 0.5, dpsd = 0, alternative = "greater")
 #' ptbf01(k = 6, n = 146, dpm = 0, dpsd = 0, alternative = "greater",
 #'        lower.tail = FALSE)
 #'
 #' ## two-sided
 #' ptbf01(k = 1/6, n = 146, dpm = 0.5, dpsd = 0)
-#' ptbf01(k = 60, n = 150, dpm = 0.5, dpsd = 0.1, lower.tail = TRUE)
+#' ptbf01(k = 6, n = 146, dpm = 0, dpsd = 0, lower.tail = FALSE)
 #'
 #' ## one-sample test
-#' ptbf01(k = 1/6, n = 5000, dpm = -0.25, dpsd = 0, alternative = "less", type = "one.sample")
+#' ptbf01(k = 1/6, n = 146, dpm = 0.5, dpsd = 0, alternative = "greater", type = "one.sample")
 #'
 #' @export
 ptbf01 <- Vectorize(FUN = ptbf01.,
@@ -256,4 +297,4 @@ ptbf01 <- Vectorize(FUN = ptbf01.,
 ## })
 ## mean(bfs < 1/10)
 ## ptbf01(k = 1/10, n1 = n1, n2 = n2, pscale = r, plocation = plocation, dpm = dpm,
-##        dpsd = dpsd, type = "two.sample", alternative = "two.sided")
+##        dpsd = dpsd, type = "two.sample", alternative = "two.sided", drange = "adaptive")