+++ package/NEWS.md
@@ -1,4 +1,5 @@
 # bfpwr 0.1
 
 - package development started
-- new functions: `bf01`, `tbf01`, `nbf01`, `pbf01`, `ptbf01`, `powerbf01`
+- new functions: `bf01`, `tbf01`, `nmbf01`, `nbf01`, `ntbf01`, `nnmbf01`,
+  `pbf01`, `ptbf01`, `pnmbf01`, `powerbf01`, `powertbf01`, `powernmbf01`